// src/app/admin/model/product.model.ts
export class Product {
    id: number;
    affiliatewebsite: string;
    category: string;
    subcategory: string;
    size: string;
    name: string;
    origprice: string;
    offerprice: string;
    longdescription: string;
    imagename: string;
    userid: string;
    url: string;
    weblink: string;
    isactive: boolean;
    isApproved: boolean;
    companyId :number;
    entrydate: Date;
  }